fix(clang-tidy): resolve narrowing, branch-clone, and destructor warnings
Torch-cpu builds with clang-tidy -warnings-as-errors. Fix 4 issues: 1-2) static_cast<int> in autodiff_graph_optimizer.cpp for size_type narrowing 3) merge identical if/else branches in core_simd_ops.cpp 4) default destructor in persistence_pipeline.hpp

+++ src/autodiff/autodiff_graph_optimizer.cpp
@@ -64,7 +64,7 @@ public:
 
     int addNode(Node::Type type, const std::vector<int> &inputs = {})
     {
-        int id = nodes.size();
+        int id = static_cast<int>(nodes.size());
         auto node = std::make_unique<Node>(id, type);
         node->inputs = inputs;
 
@@ -201,7 +201,7 @@ public:
 
         for (const auto &node : graph.nodes)
         {
-            in_degree[node->id] = node->inputs.size();
+            in_degree[node->id] = static_cast<int>(node->inputs.size());
         }
 
         std::queue<int> ready;
